So I got prismatic working locally, but I am having an issue getting it running in docker. I am working on adjusting the settings.py file to check the OS to see if its containerized and then choosing the .env var accordingly. Line 78 is where the logic takes place, but I think this code is bunk, its from an LLM and Im pretty sure it doesn't even check the OS.
If I can't get this to work I think the second solution would be to find where I defined the database parameters for the dockerized db and change them to match to the local setup.

Removing the need to do a check seems like the correct thing to do anyways. By adding a check I am creating spaghetti code. Making sure the modules work independently is key to saying functionally dynamic.